requirements.txt
asgiref==3.7.2
Django==3.2.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0

djangorestframework~=3.15.1